util: expose toUSVString
Expose toUSVString so it can be used by user libraries.

PR-URL: https://github.com/nodejs/node/pull/39814
Refs: https://github.com/nodejs/undici/pull/986
