+++ lib/internal/util.js
@@ -17,6 +17,7 @@ const {
   Promise,
   ReflectApply,
   ReflectConstruct,
+  RegExpPrototypeExec,
   RegExpPrototypeTest,
   SafeMap,
   SafeSet,
@@ -26,6 +27,10 @@ const {
   Symbol,
   SymbolFor,
 } = primordials;
+
+const {
+  toUSVString: _toUSVString,
+} = internalBinding('url');
 
 const {
   hideStackFrames,
@@ -51,6 +56,18 @@ const noCrypto = !process.versions.openssl;
 const experimentalWarnings = new SafeSet();
 
 const colorRegExp = /\u001b\[\d\d?m/g; // eslint-disable-line no-control-regex
+
+const unpairedSurrogateRe =
+  /(?:[^\uD800-\uDBFF]|^)[\uDC00-\uDFFF]|[\uD800-\uDBFF](?![\uDC00-\uDFFF])/;
+function toUSVString(val) {
+  const str = `${val}`;
+  // As of V8 5.5, `str.search()` (and `unpairedSurrogateRe[@@search]()`) are
+  // slower than `unpairedSurrogateRe.exec()`.
+  const match = RegExpPrototypeExec(unpairedSurrogateRe, str);
+  if (!match)
+    return str;
+  return _toUSVString(str, match.index);
+}
 
 let uvBinding;
 
@@ -502,6 +519,7 @@ module.exports = {
   sleep,
   spliceOne,
   structuredClone,
+  toUSVString,
   removeColors,
 
   // Symbol used to customize promisify conversion
